Sequelize and RDS Postgres Client changes

+++ server/controllers/filesController.js
@@ -12,12 +12,20 @@ const moment = require('moment');
 const md5File = require('md5-file');
 const fs = require('fs');
 const uuidv4 = require('uuid/v4');
+const aws = require('aws-sdk');
+aws.config.update({
+    "region": 'us-east-1'
+});
 
 moment.suppressDeprecationWarnings = true;
 
 const path = require('path');
 
 const multer = require('multer');
+const multerS3 = require('multer-s3');
+const s3 = new aws.S3({ apiVersion: '2006-03-01' });
+const bucket = process.env.S3_BUCKET;
+// const bucket = "to run locally uncomment this!";
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
         cb(null, 'public/uploads')
@@ -26,6 +34,29 @@ const storage = multer.diskStorage({
         cb(null, Date.now() + '-' + file.originalname)
     }
 });
+
+const uploadS3 = multer({
+    fileFilter: function (req, file, callback) {
+        var ext = path.extname(file.originalname);
+        if (ext !== '.png' && ext !== '.jpg' && ext !== '.pdf' && ext !== '.jpeg') {
+            return callback({ "Error": "Only pdfs & images are allowed" }, false);
+        }
+        callback(null, true)
+    },
+    limits: {
+        fileSize: 1024 * 1024
+    },
+    storage: multerS3({
+        s3: s3,
+        bucket: bucket,
+        acl: 'private',
+        contentType: multerS3.AUTO_CONTENT_TYPE,
+        //   serverSideEncryption: 'AES256',
+        key: function (req, file, cb) {
+            cb(null, Date.now() + '-' + file.originalname)
+        }
+    })
+}).single('billAttachment');
 
 const upload = multer({
     storage: storage,
@@ -85,49 +116,48 @@ module.exports = {
                     }
                     else {
 
-                        upload(req, res, function (err) {
+                        console.log("----------- REGION  ------------ " + aws.config.region);
+                        uploadS3(req, res, function (err) {
                             if (err) {
                                 return res.status(400).send(err);
                             } else {
-                                md5File(`public/uploads/${req.file.filename}`, (err10, hash) => {
-                                    if (err10) {
+                                console.log("-----------------------------------------------------------------------")
+                                console.log(req.file);
+                                
+                                return File
+                                    .create({
+                                        id: uuidv4(),
+                                        file_name: req.file.originalname,
+                                        url: req.file.location,
+                                        upload_date: new Date(),
+                                        size: req.file.size,
+                                        fileOwner: user.dataValues.email_address,
+                                        bill: bills[0].dataValues.id,
+                                        md5: "fieldName =>"+ req.file.fieldname,
+                                        key: req.file.key
+                                    })
+                                    .then((file) => {
+                                        delete file.dataValues.createdAt;
+                                        delete file.dataValues.updatedAt;
+                                        delete file.dataValues.fileOwner;
+                                        delete file.dataValues.size;
+                                        delete file.dataValues.bill;
+                                        delete file.dataValues.md5;
+                                        delete file.dataValues.key;
+                                        Bill
+                                            .update(
+                                                { attachment: file.dataValues.id },
+                                                {
+                                                    where: {
+                                                        id: req.params.id
+                                                    }
+                                                }
+                                            )
+                                        res.status(201).send(file);
+                                    })
+                                    .catch((error) => {
                                         res.status(400).send(error);
-                                    }
-                                    else {
-                                        return File
-                                            .create({
-                                                id: uuidv4(),
-                                                file_name: req.file.filename,
-                                                url: `public/uploads/${req.file.filename}`,
-                                                upload_date: new Date(),
-                                                size: req.file.size,
-                                                fileOwner: user.dataValues.email_address,
-                                                bill: bills[0].dataValues.id,
-                                                md5: hash
-                                            })
-                                            .then((file) => {
-                                                delete file.dataValues.createdAt;
-                                                delete file.dataValues.updatedAt;
-                                                delete file.dataValues.fileOwner;
-                                                delete file.dataValues.size;
-                                                delete file.dataValues.bill;
-                                                delete file.dataValues.md5;
-                                                Bill
-                                                    .update(
-                                                        { attachment: file.dataValues.id },
-                                                        {
-                                                            where: {
-                                                                id: req.params.id
-                                                            }
-                                                        }
-                                                    )
-                                                res.status(201).send(file);
-                                            })
-                                            .catch((error) => {
-                                                res.status(400).send(error);
-                                            });
-                                    }
-                                })
+                                    });
                             }
                         });
                     }
@@ -204,6 +234,7 @@ module.exports = {
                                 delete file[0].dataValues.size;
                                 delete file[0].dataValues.bill;
                                 delete file[0].dataValues.md5;
+                                delete file[0].dataValues.key;
                                 res.status(200).send(file[0]);
                             })
                             .catch((error) => {
@@ -290,23 +321,30 @@ module.exports = {
                                                 message: "File for this Bill Not Found!"
                                             })
                                         }
-
-                                        fs.unlink(file[0].dataValues.url, function (err) {
-
-                                            return File
-                                                .destroy({
-                                                    where: {
-                                                        id: req.params.fileId
-                                                    }
+                                        s3.deleteObject({
+                                            Bucket: bucket,
+                                            Key: file[0].key
+                                        }, function (err09) {
+                                            if (err09) {
+                                                return res.status(400).send({
+                                                    message: "Error while deleting from S3!"
                                                 })
-                                                .then((rowDeleted) => {
-                                                    if (rowDeleted === 1) {
-                                                        res.status(204).send('Deleted successfully');
-                                                    }
-                                                })
-                                                .catch((error2) => {
-                                                    res.status(400).send(error2);
-                                                });
+                                            } else {
+                                                return File
+                                                    .destroy({
+                                                        where: {
+                                                            id: req.params.fileId
+                                                        }
+                                                    })
+                                                    .then((rowDeleted) => {
+                                                        if (rowDeleted === 1) {
+                                                            res.status(204).send('Deleted successfully');
+                                                        }
+                                                    })
+                                                    .catch((error2) => {
+                                                        res.status(400).send(error2);
+                                                    });
+                                            }
                                         });
                                     })
                                     .catch((error) => {
